+++ addons/dialogic/Editor/SettingsEditor/SettingsEditor.gd
@@ -1,6 +1,8 @@
 tool
 extends ScrollContainer
 
+var editor_reference
+
 onready var nodes = {
 	'themes': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer/HBoxContainer/ThemeOptionButton,
 	'advanced_themes': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer/HBoxContainer2/AdvancedThemes,
@@ -10,16 +12,21 @@ onready var nodes = {
 	'auto_color_names': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer2/HBoxContainer3/AutoColorNames,
 	'propagate_input': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer2/HBoxContainer4/PropagateInput,
 	'dim_characters': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer2/HBoxContainer5/DimCharacters,
+	'text_event_audio_enable': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer2/HBoxContainer7/EnableVoices,
+	'text_event_audio_default_bus' : $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer2/TextAudioDefaultBus/AudioBus,
 	'save_current_timeline': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer3/HBoxContainer/SaveCurrentTimeline,
 	'clear_current_timeline': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer3/HBoxContainer2/ClearCurrentTimeline,
 	'save_definitions_on_start': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer3/HBoxContainer3/SaveDefinitionsOnStart,
 	'save_definitions_on_end': $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer3/HBoxContainer4/SaveDefinitionsOnEnd,
 	'delay_after_options': $VBoxContainer/HBoxContainer3/VBoxContainer2/VBoxContainer/HBoxContainer/LineEdit,
 	'default_action_key': $VBoxContainer/HBoxContainer3/VBoxContainer2/VBoxContainer/HBoxContainer2/DefaultActionKey,
-}
+	'canvas_layer' : $VBoxContainer/HBoxContainer3/VBoxContainer/VBoxContainer/HBoxContainer3/CanvasLayer,
+
+	'use_custom_events':$VBoxContainer/HBoxContainer3/VBoxContainer2/TimelineSection/CustomEvents/CustomEvents}
 
 var THEME_KEYS := [
 	'advanced_themes',
+	'canvas_layer',
 	]
 
 var INPUT_KEYS := [
@@ -34,6 +41,7 @@ var DIALOG_KEYS := [
 	'auto_color_names',
 	'propagate_input',
 	'dim_characters',
+	'text_event_audio_enable',
 	]
 
 var SAVING_KEYS := [
@@ -43,7 +51,13 @@ var SAVING_KEYS := [
 	'save_definitions_on_end',
 	]
 
+var EDITOR_KEYS := [
+	'use_custom_events'
+]
+
 func _ready():
+	update_bus_selector()
+	
 	update_data()
 	
 	# Themes
@@ -51,24 +65,32 @@ func _ready():
 	nodes['delay_after_options'].connect('text_changed', self, '_on_delay_options_text_changed')
 	# TODO move to theme section later
 	nodes['advanced_themes'].connect('toggled', self, '_on_item_toggled', ['dialog', 'advanced_themes'])
+	nodes['canvas_layer'].connect('text_changed', self, '_on_canvas_layer_text_changed')
 
 	nodes['default_action_key'].connect('pressed', self, '_on_default_action_key_presssed')
 	nodes['default_action_key'].connect('item_selected', self, '_on_default_action_key_item_selected')
-		
+	
+	AudioServer.connect("bus_layout_changed", self, "update_bus_selector")
+	nodes['text_event_audio_default_bus'].connect('item_selected', self, '_on_text_audio_default_bus_item_selected')
+	
 	for k in DIALOG_KEYS:
 		nodes[k].connect('toggled', self, '_on_item_toggled', ['dialog', k])
 	
 	for k in SAVING_KEYS:
 		nodes[k].connect('toggled', self, '_on_item_toggled', ['saving', k])
 
+	for k in EDITOR_KEYS:
+		nodes[k].connect('toggled', self, '_on_item_toggled', ['editor', k])
 
 func update_data():
 	var settings = DialogicResources.get_settings_config()
+	nodes['canvas_layer'].text = settings.get_value("theme", "canvas_layer", '1')
 	refresh_themes(settings)
 	load_values(settings, "dialog", DIALOG_KEYS)
 	load_values(settings, "saving", SAVING_KEYS)
 	load_values(settings, "input", INPUT_KEYS)
-
+	load_values(settings, 'editor', EDITOR_KEYS)
+	select_bus(settings.get_value("dialog", 'text_event_audio_default_bus', "Master"))
 
 func load_values(settings: ConfigFile, section: String, key: Array):
 	for k in key:
@@ -79,7 +101,7 @@ func load_values(settings: ConfigFile, section: String, key: Array):
 				if k == 'default_action_key':
 					nodes['default_action_key'].text = settings.get_value(section, k)
 				else:
-					nodes[k].pressed = settings.get_value(section, k)
+					nodes[k].pressed = settings.get_value(section, k, false)
 
 
 func refresh_themes(settings: ConfigFile):
@@ -133,12 +155,56 @@ func _on_default_action_key_presssed() -> void:
 
 
 func _on_default_action_key_item_selected(index) -> void:
-	print(index)
-	if index == 0:
-		print('here')
 	set_value('input', 'default_action_key', nodes['default_action_key'].text)
+
+
+func _on_canvas_layer_text_changed(text) -> void:
+	set_value('theme', 'canvas_layer', text)
 
 
 # Reading and saving data to the settings file
 func set_value(section, key, value):
 	DialogicResources.set_settings_value(section, key, value)
+
+func update_bus_selector():
+	if nodes["text_event_audio_default_bus"] != null:
+		var previous_selected_bus_name = ""
+		if nodes["text_event_audio_default_bus"].get_item_count():
+			previous_selected_bus_name = nodes["text_event_audio_default_bus"].get_item_text(max(0, nodes["text_event_audio_default_bus"].selected))
+
+		nodes["text_event_audio_default_bus"].clear()
+		for i in range(AudioServer.bus_count):
+			var bus_name = AudioServer.get_bus_name(i)
+			nodes["text_event_audio_default_bus"].add_item(bus_name)
+
+			if previous_selected_bus_name == bus_name:
+				nodes["text_event_audio_default_bus"].select(i)
+
+
+func select_bus(text):
+	for item_idx in range(nodes["text_event_audio_default_bus"].get_item_count()):
+		if nodes["text_event_audio_default_bus"].get_item_text(item_idx) == text:
+			nodes["text_event_audio_default_bus"].select(item_idx)
+			return
+	nodes["text_event_audio_default_bus"].select(0)
+
+
+func _on_text_audio_default_bus_item_selected(index):
+	var text = nodes['text_event_audio_default_bus'].get_item_text(index)
+	set_value('dialog', 'text_event_audio_default_bus', text)
+
+
+func _on_CustomEventsFolder_pressed():
+	editor_reference.godot_dialog("", EditorFileDialog.MODE_OPEN_DIR)
+	editor_reference.godot_dialog_connect(self, "_on_CustomEventsFolder_selected", "dir_selected")
+	#editor_reference.godot_dialog_connect(self, "_on_CustomEventsFolder_selected", "file_selected")
+
+func _on_CustomEventsFolder_selected(path, target):
+	DialogicResources.set_settings_value("editor", 'custom_events_path', path)
+	nodes['custom_events_folder_button'].text = DialogicResources.get_filename_from_path(path)
+	editor_reference.get_node("MainPanel/TimelineEditor").update_custom_events()
+
+
+
+func _on_RefreshCustomEvents_pressed():
+	editor_reference.get_node("MainPanel/TimelineEditor").update_custom_events()