+++ addons/dialogic/dialogic.gd
@@ -3,6 +3,8 @@ extends EditorPlugin
 
 var _editor_view
 var _parts_inspector
+var _export_plugin
+
 
 func _init():
 	if Engine.editor_hint:
@@ -15,10 +17,11 @@ func _init():
 		DialogicUtil.resource_fixer()
 	
 
-
 func _enter_tree() -> void:
 	_parts_inspector = load("res://addons/dialogic/Other/inspector_timeline_picker.gd").new()
 	add_inspector_plugin(_parts_inspector)
+	_export_plugin = load("res://addons/dialogic/Other/export_plugin.gd").new()
+	add_export_plugin(_export_plugin)
 	_add_custom_editor_view()
 	get_editor_interface().get_editor_viewport().add_child(_editor_view)
 	_editor_view.editor_interface = get_editor_interface()
@@ -33,10 +36,10 @@ func _ready():
 		get_editor_interface().get_resource_filesystem().scan()
 	
 
-
 func _exit_tree() -> void:
 	_remove_custom_editor_view()
 	remove_inspector_plugin(_parts_inspector)
+	remove_export_plugin(_export_plugin)
 
 
 func has_main_screen():
@@ -68,3 +71,4 @@ func _remove_custom_editor_view():
 	if _editor_view:
 		remove_control_from_bottom_panel(_editor_view)
 		_editor_view.queue_free()
+