+++ addons/dialogic/Documentation/Content/Changelog.md
@@ -1,12 +1,35 @@
+# Changelog
+
+## v1.3 - (WIP)
+- ⭐ Builtin Documentation [[Jowan-Spooner]](https://github.com/Jowan-Spooner)
+- Fixed a bug that prevented to use the global input setting when selecting option buttons
+- The TimelineEditor now paritially supports undo and redo functionality. [[Jowan-Spooner]](https://github.com/Jowan-Spooner)
+- `Dialogic.start()` will now use paths for specific timelines. A fallback is in place, but specificity is preferred. For instance: `Dialogic.start('my-timeline')` will search any timeline with that name; `Dialogic.start('/chapter-1/my-timeline')` will open the timeline namde `my-timeline` inside the folder `chapter-1`.
+- Rework of the character name coloring (using Regex now) [[Jowan-Spooner](https://github.com/Jowan-Spooner)]
+- Change the Selected Event Style to only have blue borders and not change event color [[Jowan-Spooner](https://github.com/Jowan-Spooner)]
+- Adding an extra space at the end of the timelines so it is easier to drag and drop new events
+- Fixed a reference bug that prevent the duplication of Themes
+- Added a preview image on the portrait picker, so it is easy to know what sprite or scene you are selecting. Thanks to [[EmmaH]](https://www.youtube.com/channel/UC4y59CMiLxWQQVqVFBYLa3Q) for the idea and [[Jowan-Spooner](https://github.com/Jowan-Spooner)] for the implementation
+- Fixing requirement to manually configure export dialog [[LuRomao]](https://github.com/LuRomao)
+- Changed the default cursor shape that was weird in IOS
+- Theme Editor:
+  - You can now set make the buttons be aligned horizontally or vertically
+
+
 ## v1.2.5 - Possibly breaking eveything. We will never know.
 - Loading timeline events on batches to speedup big timelines
 - Reduced the amount of nodes inside events to improve loading times
 - Fixing the "first time running" bug where you had to reboot dialogic after enabling it for the first time
 - Fixing a bug where if you had a node selected in the editor you couldn't open any resource picker
 - Fixed a bug when trying to add a new resource after removing one
+- Fixed the issue that allowed you to keep loading messages after a dialog close event
+- Enabling bbcode to glossary entries
 - Simplified some internal code
+- Adding new setting to the Set Background event to add fade-in time
+- Fixed an issue that caused nested timelines to be deleted whem moving directories [[zaknafean]](https://github.com/zaknafean)
 - Settings Editor
   - Added a default action key selector so you don't have to set it per theme. The theme action key settings will overwrite the one set in settings. 
+  - Added new setting to select Dialogic's Canvas Layer [[RedXGames]](https://github.com/RedXGames)
 
 ## v1.2.4 - Gotta go fast
 - Fixed an issue with the default scale of the portraits